+++ source/config/Options.ts
@@ -50,7 +50,7 @@ export class Options {
 
     {
       brand: OptionBrand.Boolean,
-      description: "Check error messages provided with the '// @ts-expect-error' directives.",
+      description: "Check errors silenced by '// @ts-expect-error' directives.",
       group: OptionGroup.ConfigFile,
       name: "checkSuppressedErrors",
     },