+++ source/select/Select.ts
@@ -2,8 +2,8 @@ import fs from "node:fs/promises";
 import type { ResolvedConfig } from "#config";
 import { Diagnostic } from "#diagnostic";
 import { EventEmitter } from "#events";
+import { Glob } from "#glob";
 import { Path } from "#path";
-import { GlobPattern } from "./GlobPattern.js";
 import { SelectDiagnosticText } from "./SelectDiagnosticText.js";
 
 interface FileSystemEntryMeta {
@@ -52,8 +52,8 @@ export class Select {
 
     if (!matchPatterns) {
       matchPatterns = {
-        includedDirectory: GlobPattern.toRegex(globPatterns, "directories"),
-        includedFile: GlobPattern.toRegex(globPatterns, "files"),
+        includedDirectory: Glob.toRegex(globPatterns, "directories"),
+        includedFile: Glob.toRegex(globPatterns, "files"),
       };
 
       Select.#patternsCache.set(globPatterns, matchPatterns);