+++ tests/__utilities__/assert.js
@@ -5,6 +5,8 @@ import process from "node:process";
 
 export { doesNotMatch, equal, match, notEqual } from "node:assert/strict";
 
+// TODO Consider 'assert.partialDeepStrictEqual()' after dropping support for Node.js 20
+
 /**
  * @param {Record<string, unknown> | string} source
  * @param {Record<string, unknown>} target